pyproject.toml
[project]
name = "agent-kit-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi[standard]>=0.119.0",
    "openai-agents>=0.4.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]
